Add spatial support and docs

+++ R/downloads.R
@@ -1,8 +1,58 @@
-download_sample <- function(sample_id, auth_token, path = "scpca_data", format = "sce") {
+#' Download a sample's data files from the ScPCA Portal
+#'
+#' This function downloads the data files for a specified sample from the ScPCA Portal.
+#' The downloaded files are saved in a subdirectory of the specified path, named
+#' by the base filename of the downloaded zip file, which includes the sample ID, modality,
+#' format, and date.
+#'
+#' For single-cell and single-nucleotide data, files can be downloaded in either
+#' SingleCellExperiment ("sce") or AnnData ("anndata") format,
+#' and all downloads include the unfiltered, filtered, and processed data objects, as well as
+#' associated metadata and QC files.
+#'
+#' Spatial data, if present, can be downloaded in Space Ranger format using the "spatial" format option.
+#'
+#' The function returns a vector of file paths for the downloaded files (invisibly).
+#'
+#' Note that downloading data requires an authorization token, which can be obtained
+#' using the `get_auth()` function.
+#'
+#'
+#' @param sample_id The SCPCA sample ID (e.g. "SCPCS000001")
+#' @param auth_token An authorization token obtained from `get_auth()`
+#' @param path The path to the directory where the unzipped file directory should be saved. Default is "scpca_data".
+#' @param format The desired file format, either "sce" (SingleCellExperiment),
+#'  "anndata" (AnnData/H5AD), or "spatial" (for spatial data in Space Ranger format).
+#'  Default is "sce".
+#' @param overwrite Whether to overwrite existing directories if they already exist. Default is FALSE.
+#' @param quiet Whether to suppress download progress messages. Default is FALSE.
+#'
+#' @import httr2
+#'
+#' @returns a vector of file paths for the downloaded files (invisibly)
+#'
+#' @export
+#' @examples
+#' \dontrun{
+#' # Get a token first
+#' auth_token <- get_auth("me@email.net", agree = TRUE)
+#' # Then ask for a sample download
+#' download_sample("SCPCS000001", auth_token, path = "scpca_data", format = "sce")
+#'
+#' # Downloading in AnnData format
+#' download_sample("SCPCS000001", auth_token, path = "scpca_data", format = "anndata")
+#' }
+download_sample <- function(
+    sample_id,
+    auth_token,
+    path = "scpca_data",
+    format = "sce",
+    overwrite = FALSE,
+    quiet = FALSE) {
   stopifnot(
     "Invalid sample_id" = grepl("^SCPCS\\d{6}$", sample_id),
-    "path must be a valid directory" = dir.exists(path),
-    "Authorization token must be provided" = is.character(auth_token) && nchar(auth_token) > 0
+    "Authorization token must be provided" = is.character(auth_token) && nchar(auth_token) > 0,
+    "quiet must be a logical value" = is.logical(quiet) && length(quiet) == 1
   )
 
   sce_formats <- c(
@@ -12,13 +62,25 @@ download_sample <- function(sample_id, auth_token, path = "scpca_data", format =
     "single_cell_experiment"
   )
   anndata_formats <- c("anndata", "h5ad")
+  spatial_formats <- c("spatial", "spaceranger")
 
-  if (tolower(format) %in% sce_formats) {
+  # normalize format input to match API values
+  format <- tolower(format)
+  if (format %in% sce_formats) {
     format_str <- "SINGLE_CELL_EXPERIMENT"
-  } else if (tolower(format) %in% anndata_formats) {
+  } else if (format %in% anndata_formats) {
     format_str <- "ANN_DATA"
+  } else if (format %in% spatial_formats) {
+    format_str <- "SPATIAL"
   } else {
-    stop("Invalid format. Supported formats are: 'sce', 'anndata', 'h5ad'.")
+    stop(
+      "Invalid format. Expected format strings are 'sce', 'anndata', or 'spatial'",
+      " (with some additional variants accepted)."
+    )
+  }
+  # create target directory if it doesn't exist
+  if (!dir.exists(path)) {
+    dir.create(path, recursive = TRUE)
   }
 
   sample_info <- tryCatch(
@@ -35,15 +97,28 @@ download_sample <- function(sample_id, auth_token, path = "scpca_data", format =
     stop(glue::glue("Sample `{sample_id}` not found."))
   }
 
+  # message if multiplexed
+  if (sample_info$has_multiplexed_data) {
+    multiplexed_ids <- sample_info$multiplex |>
+      paste(collapse = ", ")
+    message(glue::glue(
+      "Sample {sample_id} is multiplexed with other samples.",
+      " Downloading all associated libraries.",
+      " Note that directory names will include all multiplexed sample IDs."
+    ))
+  }
+
   file_list <- sample_info$computed_files |>
-    purrr::keep(\(f) f$format == format_str)
+    # filter to requested format or modality (only applies to spatial data)
+    purrr::keep(\(cf) {
+      (cf$format == format_str || cf$modality == format_str)
+    })
   if (length(file_list) == 0) {
     stop(glue::glue("No files found for sample {sample_id} in format {format}."))
   }
 
   # build requests for each file
-  # there really should only be one file per sample/format, but just in case
-  # I might be missing something about spatial data
+  # most samples will only have one file, but in some multiplexed cases there may be more
   file_requests <- file_list |>
     purrr::map_chr(as.character("id")) |>
     purrr::map(\(id) {
@@ -57,19 +132,36 @@ download_sample <- function(sample_id, auth_token, path = "scpca_data", format =
   download_urls <- req_perform_parallel(file_requests) |>
     resps_data(\(resp) resp_body_json(resp)$download_url)
 
-  file_paths <- purrr::map_chr(download_urls, \(url) {
-    base_dir <- extract_download_basename(url)
-    download_dir <- file.path(path, base_dir)
+  file_paths <- purrr::map(download_urls, \(url) {
+    download_filename <- parse_download_file(url)
+    destination_dir <- file.path(path, stringr::str_remove(download_filename, "\\.zip$"))
 
-    file_temp <- tempfile(pattern = "scpca_download", fileext = ".zip")
+    # exit if directory already exists
+    if (dir.exists(destination_dir) && !overwrite) {
+      message(glue::glue(
+        "Directory {destination_dir} already exists; skipping download.",
+        " Use 'overwrite = TRUE' to replace existing files."
+      ))
+      # no files downloaded, return empty vector
+      return(c())
+    }
+
+    # TODO: Do we want to warn if a directory exists that matches except for the date?
+
+    file_temp <- file.path(tempdir(), download_filename)
     on.exit(unlink(file_temp), add = TRUE)
 
-    message(glue::glue("Downloading {base_dir} ..."))
-    curl::curl_download(url, file_temp, quiet = FALSE)
+    if (!quiet) {
+      message(glue::glue("Downloading {download_filename} ..."))
+    }
+    curl::curl_download(url, file_temp, quiet = quiet)
 
-    message(glue::glue("Unzipping to {download_dir}..."))
-    unzip(file_temp, exdir = download_dir)
-  })
+    if (!quiet) {
+      message(glue::glue("Unzipping to {destination_dir}..."))
+    }
+    utils::unzip(file_temp, exdir = destination_dir)
+  }) |>
+    purrr::list_c()
   invisible(file_paths)
 }
 
@@ -77,11 +169,11 @@ download_sample <- function(sample_id, auth_token, path = "scpca_data", format =
 #'
 #' (this may become obsolete if we get download filenames in the API response)
 #'
-#' @param url
+#' @param scpca_url The ScPCA portal download URL
 #'
-#' @returns the base filename without the .zip extension
-extract_download_basename <- function(url) {
-  params <- curl::curl_parse_url(file_url)$params
+#' @returns the download filename
+parse_download_file <- function(scpca_url) {
+  params <- curl::curl_parse_url(scpca_url)$params
   params["response-content-disposition"] |>
-    stringr::str_extract("SCPC[^\\s]+(?=\\.zip)")
+    stringr::str_extract("SCPC[^\\s]+\\.zip")
 }
